% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{function to retry an iteration of a for loop if there is an error.}
\usage{
retry(expr, isError = function(x) "try-error" \%in\% class(x),
  maxErrors = 5, sleep = 0)
}
\arguments{
\item{expr}{is the expression to retry}

\item{maxErrors}{is the number of retrys that will occur before moving on to the next iteration}

\item{sleep}{is the lag time between retrys}
}
\description{
function to retry an iteration of a for loop if there is an error.
}
\examples{

x <- c(1,2,'z',4,5,6,7)

for(i in 1:length(x)){
retry(expr = {print(runif(x[i]))}, maxErrors = 3, sleep = 1)
}


}
